le, examples: Use single precision floating point values by default

+++ examples/logistic-regression.c
@@ -5,16 +5,16 @@
 int
 main(int argc, const char *argv[])
 {
-    const double x_data[] =
+    const float x_data[] =
     {
-        1.0, 2.0, 3.0, 4.0,
-        4.0, 3.0, 2.0, 1.0
+        1.0f, 2.0f, 3.0f, 4.0f,
+        4.0f, 3.0f, 2.0f, 1.0f
     };
     LeMatrix *x = le_matrix_new_from_data(2, 4, x_data);
     
-    const double y_data[] =
+    const float y_data[] =
     {
-        0.0, 0.0, 1.0, 1.0
+        0.0f, 0.0f, 1.0f, 1.0f
     };
     LeMatrix *y = le_matrix_new_from_data(1, 4, y_data);
